author(ca)	R Documentation

Author dataset

Description

This data matrix contains the counts of the 26 letters of the alphabet (columns of matrix) for 12 different novels (rows of matrix). Each row contains letter counts in a sample of text from each work, excluding proper nouns.

Usage

data(author)

Format

Data frame containing the 12 x 26 matrix.

Source

W.A. Larsen & R. McGill, unpublished data collected in 1973.

[Package ca version 0.1-8 Index]