ca(ca)	R Documentation

Simple correspondence analysis

Description

Computation of simple correspondence analysis.

Usage

ca(obj, nd = NA, suprow = NA, supcol = NA, 
   subsetrow = NA, subsetcol = NA)

Arguments

obj 	 A two-way table of non-negative data, usually frequencies.

nd 	 Number of dimensions to be included in the output; if NA the maximum possible dimensions are included.

suprow 	 Indices of supplementary rows.

supcol 	 Indices of supplementary columns.

subsetrow 	 Row indices of subset.

subsetcol 	 Column indices of subset.

Details

The function ca computes a simple correspondence analysis based on the singular value decomposition.
The options suprow and supcol allow supplementary (passive) rows and columns to be specified. Using the options subsetrow and/or subsetcol result in a subset CA being performed.

Value

sv 	 Singular values

nd 	 Dimenson of the solution

rownames 	 Row names

rowmass 	 Row masses

rowdist 	 Row chi-square distances to centroid

rowinertia 	 Row inertias

rowcoord 	 Row standard coordinates

rowsup 	 Indices of row supplementary points

colnames 	 Column names

colmass 	 Column masses

coldist 	 Column chi-square distances to centroid

colinertia 	 Column inertias

colcoord 	 Column standard coordinates

colsup 	 Indices of column supplementary points

References

Nenadic, O. and Greenacre, M. (2007). Correspondence analysis in R, with two- and three-dimensional graphics: The ca package. Journal of Statistical Software, 20 (3), available at http://www.jstatsoft.org/v20/i03/
Greenacre, M. (2007). Correspondence Analysis in Practice. Second Edition. Chapman & Hall / CRC, London.
Blasius, J. and Greenacre, M. J. (1994), Computation of correspondence analysis, in ``Correspondence Analysis in the Social Sciences'', pp. 53-75, Academic Press, London.
Greenacre, M.J. and Pardo, R. (2006), Subset correspondence analysis: visualizing relationships among a selected set of response categories from a questionnaire survey. Sociological Methods and Research, 35, pp. 193-218.

See Also

svd, plot.ca, plot3d.ca, summary.ca, print.ca

Examples

 
data(author)
ca(author)
plot(ca(author))
 

[Package ca version 0.33 Index]