iterate.mjca(ca)	R Documentation

Updating a Burt matrix in Joint Correspondence Analysis

Description

Updating a Burt matrix in Joint Correspondence Analysis based on iteratively weighted least squares.

Usage

iterate.mjca(B, lev.n, nd = 2, maxit = 50, epsilon = 0.0001)

Arguments

B 	 A Burt matrix.

lev.n 	 The number of levels for each factor from the original response pattern matrix.

nd 	 The required dimensionality of the solution.

maxit 	 The maximum number of iterations.

epsilon 	 A convergence criterion for the maximum absolute difference of updated values compared to the previous values. The iteration is completed when all differences are smaller than epsilon.

Details

The function iterate.mjca computes the updated Burt matrix. This function is called from the function mjca when the option lambda="JCA", i.e. when a Joint Correspondence Analysis is performed.

Value

B.star 	 The updated Burt matrix

crit 	 Vector of length 2 containing the number of iterations and epsilon

See Also

mjca

[Package ca version 0.3-2 Index]