mjca(ca)	R Documentation

Multiple and joint correspondence analysis

Description

Computation of multiple and joint correspondence analysis.

Usage

mjca(obj, nd = 2, lambda = "adjusted", supcol = NA, subsetcol = NA, 
     ps = "", maxit = 50, epsilon = 0.0001)

Arguments

obj 	 A response pattern matrix (data frame containing factors).

nd 	 Number of dimensions to be included in the output; if NA the maximum possible dimensions are included.

lambda 	 Gives the scaling method. Possible values include "indicator", "Burt", "adjusted" and "JCA". Using lambda = "JCA" results in a joint correspondence analysis using iterative adjusment of the Burt matrix in the solution space.

supcol 	 Indices of supplementary columns.

subsetcol 	 Indices of subset categories.

ps 	 Separator used for combining variable and category names.

maxit 	 The maximum number of iterations (Joint Correspondence Analysis).

epsilon 	 A convergence criterion (Joint Correspondence Analysis).

Details

The function mjca computes a multiple or joint correspondence analysis based on the eigenvalue decomposition of the Burt matrix.

Value

sv 	 Eigenvalues (lambda = "indicator") or singular values (lambda = "Burt", "adjusted" or "JCA")

lambda 	 Scaling method

inertia.e 	 Percentages of explained inertia

inertia.t 	 Total inertia

inertia.et 	 Total percentage of explained inertia with the nd-dimensional solution

levelnames 	 Names of the factor/level combinations

levels.n 	 Number of levels in each factor

nd 	 User-specified dimensionality of the solution

nd.max 	 Maximum possible dimensionality of the solution

rownames 	 Row names

rowmass 	 Row masses

rowdist 	 Row chi-square distances to centroid

rowinertia 	 Row inertias

rowcoord 	 Row standard coordinates

colnames 	 Column names

colmass 	 Column masses

coldist 	 Column chi-square distances to centroid

colinertia 	 Column inertias

colcoord 	 Column standard coordinates

colsup 	 Indices of column supplementary points (of the Burt and Indicator matrix)

subsetcol 	 Indices of subset columns

Burt 	 Burt matrix

Burt.upd 	 The updated Burt matrix (JCA only)

subinertia 	 Inertias of sub-matrices

JCA.iter 	 Vector of length two containing the number of iterations and the epsilon (JCA only)

call 	 Return of match.call

References

Nenadic, O. and Greenacre, M. (2007), Correspondence analysis in R, with two- and three-dimensional graphics: The ca package. Journal of Statistical Software, 20 (3), available at http://www.jstatsoft.org/v20/i03/
Nenadic, O. and Greenacre, M. (2007), Computation of Multiple Correspondence Analysis, with Code in R, in Multiple Correspondence Analysis and Related Methods (eds. M. Greenacre and J. Blasius), Chapmann & Hall / CRC, Boca Raton, London, New York, pp. 523-551.
Greenacre, M.J. and Pardo, R. (2006), Subset correspondence analysis: visualizing relationships among a selected set of response categories from a questionnaire survey. Sociological Methods and Research, 35, pp. 193-218.

See Also

eigen, plot.mjca, summary.mjca, print.mjca

Examples

 
library(MASS)
data(farms)
mjca(farms)

# Joint correspondence analysis:
mjca(farms, lambda = "JCA")

 

[Package ca version 0.33 Index]