pchlist(ca)	R Documentation

Listing the set of available symbols.

Description

A plot of the available symbols for use with the option pch.

Usage

pchlist() 

Details

This function generates a numbered list of the plotting symbols available for use in the functions plot.ca and plot3d.ca.

See Also

plot.ca, plot3d.ca

Examples


pchlist()

[Package ca version 0.3-2 Index]