plot3d.ca(ca)	R Documentation

Plotting 3D maps in correspondence analysis

Description

Graphical display of correspondence analysis in three dimensions

Usage

plot3d.ca(x, dim = c(1, 2, 3), map = "symmetric", what = c("all", "all"), 
             contrib = c("none", "none"), col = c("#6666FF","#FF6666"), 
             labcol  = c("#0000FF", "#FF0000"), pch = c(16, 1, 18, 9), 
             labels = c(2, 2), sf = 0.00002, arrows  = c(FALSE, FALSE), ...) 

Arguments

x 	 Simple correspondence analysis object returned by ca

dim 	 Numerical vector of length 2 indicating the dimensions to plot

map 	 Character string specifying the map type. Allowed options include
"symmetric" (default)
"rowprincipal"
"colprincipal"
"symbiplot"
"rowgab"
"colgab"
"rowgreen"
"colgreen"

what 	 Vector of two character strings specifying the contents of the plot. First entry sets the rows and the second entry the columns. Allowed values are
"none" (no points are displayed)
"active" (only active points are displayed, default)
"supplementary" (only supplementary points are displayed)
"all" (all available points)
The status (active or supplementary) is set in ca.

contrib 	 Vector of two character strings specifying if contributions (relative or absolute) should be indicated by different colour intensities. Available options are
"none" (contributions are not indicated in the plot).
"absolute" (absolute contributions are indicated by colour intensities).
"relative" (relative conrributions are indicated by colour intensities).
If set to "absolute" or "relative", points with zero contribution are displayed in white. The higher the contribution of a point, the closer the corresponding colour to the one specified by the col option.

col 	 Vector of length 2 specifying the colours of row and column profiles. Colours can be entered in hexadecimal (e.g. "#FF0000"), rgb (e.g. rgb(1,0,0)) values or by R-name (e.g. "red").

labcol 	 Vector of length 2 specifying the colours of row and column labels.

pch 	 Vector of length 2 giving the type of points to be used for rows and columns.

labels 	 Vector of length two specifying if the plot should contain symbols only (0), labels only (1) or both symbols and labels (2). Setting labels to 2 results in the symbols being plotted at the coordinates and the labels with an offset.

sf 	 A scaling factor for the volume of the 3d primitives.

arrows 	 Vector of two logicals specifying if the plot should contain points (FALSE, default) or arrows (TRUE). First value sets the rows and the second value sets the columns.

... 	 Further arguments passed to the rgl functions.

See Also

ca

[Package ca version 0.3-2 Index]