print.ca(ca)	R Documentation

Printing ca objects

Description

Printing method for correspondence analysis objects

Usage

## S3 method for class 'ca':
print(x, ...) 

Arguments

x 	 Simple correspondence analysis object returned by ca

... 	 Further arguments are ignored

Details

The function print.ca gives the basic statistics of the ca object. First the eigenvalues (that is, principal inertias) and their percentages with respect to total inertia are printed. Then for the rows and columns respectively, the following are printed: the masses, chi-square distances of the points to the centroid (i.e., centroid of the active points), point inertias (for active points only) and principal coordinates on the first nd dimensions requested (default = 2 dimensions). The function summary.ca gives more detailed results about the inertia contributions of each point on each principal axis.
For supplementary points, masses and inertias are not applicable.

See Also

ca

Examples


data(smoke)
print(ca(smoke))

[Package ca version 0.3-2 Index]