print.mjca(ca)	R Documentation

Printing mjca objects

Description

Printing method for multiple and joint correspondence analysis objects

Usage

## S3 method for class 'mjca':
print(x, ...) 

Arguments

x 	 Multiple or joint correspondence analysis object returned by mjca

... 	 Further arguments are ignored

Details

The function print.mjca gives the basic statistics of the mjca object. First the eigenvalues (that is, principal inertias) and their percentages with respect to total inertia are printed. Then for the rows and columns respectively, the following are printed: the masses, chi-square distances of the points to the centroid (i.e., centroid of the active points), point inertias (for active points only) and principal coordinates on the first nd dimensions requested (default = 2 dimensions). The function summary.mjca gives more detailed results about the inertia contributions of each point on each principal axis.
For supplementary points, masses and inertias are not applicable.

See Also

mjca

Examples


library(MASS)
data(farms)
print(mjca(farms))

[Package ca version 0.3-2 Index]