summary.ca(ca)	R Documentation

Summarizing simple correspondence analysis

Description

Textual output summarizing the results of ca, including a scree-plot of the principal inertias and row and column contributions.

Usage

## S3 method for class 'ca':
summary(object, scree = TRUE, ...)

Arguments

object 	 Simple correspondence analysis object returned by ca.

scree 	 Logical flag specifying if a scree-plot should be included in the output.

... 	 Further arguments (ignored)

Details

The function summary.ca gives the detailed numerical results of the ca function. All the eigenvalues (principal inertias) are listed, their percentages with respect to total inertia, and a bar chart (also known as a scree plot). Then for the set of rows and columns a table of results is given in a standard format, where quantities are either multiplied by 1000 or expressed in permills (thousandths): the mass of each point (x1000), the quality of display in the solution subspace of nd dimensions, the inertia of the point (in permills of the total inertia), and then for each dimension of the solution the principal coordinate (x1000), the (relative) contribution COR of the principal axis to the point inertia (x1000) and the (absolute) contribution CTR of the point to the inertia of the axis (in permills of the principal inertia).
For supplementary points, masses, inertias and absolute contributions (CTR) are not applicable, but the relative contributions (COR) are valid as well as their sum over the set of chosen nd dimensions (QLT).

Examples


data(smoke)
summary(ca(smoke))

[Package ca version 0.3-2 Index]