summary.mjca(ca)	R Documentation

Summarizing multiple and joint correspondence analysis

Description

Textual output summarizing the results of mjca, including a scree-plot of the principal inertias and row and column contributions.

Usage

## S3 method for class 'mjca':
summary(object, scree = TRUE, rows = FALSE, ...)

Arguments

object 	 Multiple or joint correspondence analysis object returned by mjca.

scree 	 Logical flag specifying if a scree-plot should be included in the output.

rows 	 Logical specifing whether the results for the rows should be included in the output (default = FALSE).

... 	 Further arguments (ignored)

Details

The function summary.mjca gives the detailed numerical results of the mjca function. All the eigenvalues (principal inertias) are listed, their percentages with respect to total inertia, and a bar chart (also known as a scree plot). Then for the set of rows and columns a table of results is given in a standard format, where quantities are either multiplied by 1000 or expressed in permills (thousandths): the mass of each point (x1000), the quality of display in the solution subspace of nd dimensions, the inertia of the point (in permills of the total inertia), and then for each dimension of the solution the principal coordinate (x1000), the (relative) contribution COR of the principal axis to the point inertia (x1000) and the (absolute) contribution CTR of the point to the inertia of the axis (in permills of the principal inertia).
For supplementary points, masses, inertias and absolute contributions (CTR) are not applicable, but the relative contributions (COR) are valid as well as their sum over the set of chosen nd dimensions (QLT).

Examples


 library(MASS)
 data(farms)
 summary(mjca(farms))

[Package ca version 0.3-2 Index]