wg93(ca)	R Documentation

International Social Survey Program on Environment 1993 - western German sample

Description

This data frame contains records of four questions on attitude towards science with responses on a five-point scale (1=agree strongly to 5=disagree strongly) and three demographic variables (sex, age and education).

Usage

data(wg93)

Format

Data frame (871x7).

Source

ISSP (1993). International Social Survey Program: Environment. http://www.issp.org

[Package ca version 0.1-8 Index]